feat: implement `react-redux` `Provider`
